% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDaPars.R
\name{scDaPars}
\alias{scDaPars}
\title{Use scDaPars to quantify APA dynamics across single cells using scRNA-seq data}
\usage{
scDaPars(raw_PDUI_file, out_dir, filter_gene_thre, filter_cell_thre)
}
\arguments{
\item{raw_PDUI_file}{A character specifying the full path of the raw PDUI matrix generated by step1 of scDaPars;}

\item{out_dir}{A character specifying the full path of the output directory;}

\item{filter_gene_thr}{A number between 0 and 1;
specifying the percent of cells a gene's PDUI must be detected;}

\item{filter_cell_thr}{A number between 0;
specigying the percent of genes' PDUI a cell must be detected;}
}
\value{
scDaPars returns the imputed PDUI_matrix with rows representing genes and columns representing cells;
It saves the imputed PDUI matrix as scDaPars_imputed_results.txt;
}
\description{
Use scDaPars to quantify APA dynamics across single cells using scRNA-seq data
}
\author{
Yipeng Gao, \email{yipeng.gao@bcm.edu}

Wei Li, \email{wei.li@uci.edu}
}
